requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.94.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pydantic==1.10.6
python-dotenv==1.0.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.0.post1
typing-extensions==4.5.0
urllib3==1.26.14
uvicorn==0.21.0
